+++ tests/test_main.py
@@ -671,6 +671,14 @@ def test_newline_preserved(runner, tmp_path, file, newline):
         assert output_file.newlines == newline
 
 
+def test_no_format_python_code_blocks(runner):
+    file = ".. code-block:: python\n\n" "    def example_function():\n"
+    args = ["-t", "rst", "-l", 80, "-r", file, "--no-format-python-code-blocks"]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 0
+    assert result.output == file
+
+
 @pytest.mark.parametrize("file", test_files)
 def test_globbing(runner, file):
     args = [